+++ tests/unit/apps/test_preprocessing.py
@@ -4,15 +4,13 @@ import pytest
 from docarray import Document, DocumentArray
 
 from now.app.base.preprocess import preprocess_text
-from now.app.base.transform_docarray import transform_docarray
 from now.app.search_app import SearchApp
 
 
 def test_search_app_preprocessing_query():
     """Test if the text to video preprocessing works for queries"""
     app = SearchApp()
-    da = DocumentArray([Document(text='test')])
-    da = transform_docarray(da)
+    da = DocumentArray([Document(chunks=[Document(text='test', modality='text')])])
     da = app.preprocess(da)
 
     assert len(da) == 1
@@ -24,9 +22,17 @@ def test_search_app_preprocessing_indexing(resources_folder_path):
     """Test if the text to video preprocessing works for indexing"""
     app = SearchApp()
     da = DocumentArray(
-        [Document(uri=os.path.join(resources_folder_path, 'gif/folder1/file.gif'))]
+        [
+            Document(
+                chunks=[
+                    Document(
+                        uri=os.path.join(resources_folder_path, 'gif/folder1/file.gif'),
+                        modality='video',
+                    )
+                ]
+            )
+        ]
     )
-    da = transform_docarray(da)
     da = app.preprocess(da)
     assert len(da) == 1
     assert len(da[0].chunks[0].chunks) == 3
@@ -43,8 +49,7 @@ def test_search_app_preprocessing_indexing(resources_folder_path):
 def test_text_preprocessing(app_cls, is_indexing):
     """Test if the text to text preprocessing works for queries and indexing"""
     app = app_cls()
-    da = DocumentArray([Document(text='test')])
-    da = transform_docarray(da)
+    da = DocumentArray([Document(chunks=[Document(text='test', modality='text')])])
     da = app.preprocess(da)
     assert len(da) == 1
     assert len(da[0].chunks) == 1
@@ -64,8 +69,7 @@ def test_image_preprocessing(app_cls, is_indexing, resources_folder_path):
     """Test if the image to image preprocessing works for queries and indexing"""
     app = app_cls()
     uri = os.path.join(resources_folder_path, 'image/a.jpg')
-    da = DocumentArray([Document(uri=uri)])
-    da = transform_docarray(da)
+    da = DocumentArray([Document(chunks=[Document(uri=uri, modality='image')])])
     da = app.preprocess(da)
     assert len(da) == 1
     assert len(da[0].chunks) == 1
